+++ src/pages/home/Home.css
@@ -1,4 +1,4 @@
-.calendar {
+.rightside {
   display: flex;
   flex-direction: column;
