+++ src/pages/home/right/AdminHome.css
@@ -1,0 +1,20 @@
+.event {
+  background-color: var(--purple2);
+  position: absolute;
+  z-index: 2;
+  border-radius: 10px;
+}
+
+.event {
+  background-color: var(--purple2);
+}
+.event .content {
+  width: 100%;
+  height: 100%;
+  background-color: aqua;
+  overflow: hidden;
+  padding: 7px;
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+}
